Added Metadata to auth/login
